fix and clean the code, and add portrait features
