+++ App.js
@@ -1,11 +1,12 @@
 import React from 'react';
-import { View } from 'react-native'
-import Navigation from './src/routes/Navigation'
+import Navigation from './src/routes/Navigation';
 
-export default class App extends React.Component{
+class App extends React.Component{
     render() {
-      return(
-        <Navigation/>
-      )
+        return(
+          <Navigation/>
+        );
     }
 }
+
+export default App;
