+++ package.json
@@ -10,6 +10,7 @@
     "react": "16.6.0-alpha.8af6728",
     "react-native": "0.57.4",
     "react-native-google-places": "^2.5.2",
+    "react-native-keyboard-aware-scroll-view": "^0.7.4",
     "react-native-linear-gradient": "^2.4.2",
     "react-native-maps": "^0.22.0",
     "react-navigation": "^2.18.2"
