+++ src/components/assets/fingerTouch.png
