+++ src/components/components_A/Home.js
@@ -1,7 +1,10 @@
 import React from 'react';
-import { View, Text, TouchableOpacity, Image, ImageBackground, StyleSheet } from 'react-native';
+import { View, Text, TouchableOpacity, Image, ImageBackground } from 'react-native';
+import { styles } from '../styles/Styles.Home';
+import BgHome from '../assets/BgHome.png';
+import fingerTouch from '../assets/fingerTouch.png';
 
-export default class Home extends React.Component {
+class Home extends React.Component {
     moveScreen() {
         this.props.navigation.navigate('Screen2')
     }
@@ -9,7 +12,7 @@ export default class Home extends React.Component {
         return(
                 <View style={styles.container}>
                     <ImageBackground
-                        source={require('./assets/BgHome.png')} 
+                        source={BgHome} 
                         style={styles.bg}
                     >
                     <Text style={styles.text1}>Hello!</Text>
@@ -18,14 +21,14 @@ export default class Home extends React.Component {
                             style={styles.finger}
                         >
                             <Image
-                                source={require('./assets/fingerTouch.png')}
+                                source={fingerTouch}
                                 style={styles.fingerIcon}
                             />
                         </TouchableOpacity>
                         
                         <TouchableOpacity 
                             onPress={() => this.moveScreen()}
-                            style={styles.login}
+                            style={styles.loginButton}
                         >
                             <View style={styles.loginView}>
                                 <Text style={styles.loginText}>Login</Text>
@@ -34,60 +37,8 @@ export default class Home extends React.Component {
                         
                     </ImageBackground>
                 </View>
-        )
+        );
     }
 }
 
-const styles = StyleSheet.create({
-    container:{
-        flex: 1
-    },
-    text1: {
-        color: 'white',
-        fontSize: 60,
-        fontWeight: '400',
-        fontFamily: 'HelveticaNeueLight',
-        paddingLeft: 25,
-        marginTop: 60,
-    },
-    text2: {
-        color: 'white',
-        fontSize: 40,
-        paddingLeft: 25,
-        marginTop: 10,
-        fontFamily: 'HelveticaNeueLight',
-
-    },
-    bg: { 
-        height: '100%', 
-        width: '100%' 
-    },
-    finger: {  
-        position: 'absolute', 
-        alignSelf: 'center',
-        justifyContent: 'center', 
-        marginTop: 350,
-    },
-    fingerIcon: { 
-        height: 100, 
-        width: 100,
-    },
-    login: { 
-        position: 'absolute', 
-        alignSelf: 'center', 
-        justifyContent: 'center', 
-        marginTop: 540
-    },
-    loginView: { 
-        backgroundColor: 'rgb(210, 210, 210)', 
-        height: 50, 
-        width: 300, 
-        borderRadius: 10, 
-    },
-    loginText: { 
-        fontSize: 16,
-        textAlign: 'center', 
-        fontFamily: 'HelveticaNeueLight',
-        marginTop: 15,
-    },
-})
+export default Home;
