+++ src/components/components_A/Map.js
@@ -1,19 +1,17 @@
 import React, {Component} from 'react';
-import { View, StyleSheet, Image } from 'react-native';
+import { View, Image } from 'react-native';
 import MapView from 'react-native-maps';
 import CalloutMap from '../components_B/Callout.Button';
 import CalloutDesc from '../components_B/Callout.Desc';
-
-const pin = require('./assets/pin.png')
+import pin from '../assets/pin.png';
+import {styles} from '../styles/Styles.Map';
 
 class Map extends Component{
-
     render () {
         return(
             <View style={styles.container}>
                 <MapView
                     style={styles.map}
-                    showsUserLocation={false}
                     initialRegion = {{
                         latitude: 40.7112432,
                         longitude: -74.0071598,
@@ -98,42 +96,5 @@ class Map extends Component{
         );
     }
 }
-
-const styles = StyleSheet.create({
-    container: {
-        ...StyleSheet.absoluteFillObject,
-        height: 650,
-        width: 360,
-        justifyContent: 'flex-end',
-        alignItems: 'center',
-    },
-    map: {
-        ...StyleSheet.absoluteFillObject,
-    },
-      radius: {
-        height: 80,
-        width: 80,
-        borderRadius: 80 / 2,
-        overflow: 'hidden',
-        backgroundColor: 'rgba(45, 183, 132, 0.1)',
-        borderWidth: 1,
-        borderColor: 'rgba(45, 183, 132, 0.3)',
-        alignItems: 'center',
-        justifyContent: 'center',
-    },
-      marker: {
-          height: 20,
-          width: 20,
-          borderWidth: 3,
-          borderColor: 'white',
-          borderRadius: 20 / 2,
-          overflow: 'hidden',
-          backgroundColor: 'rgb(45, 183, 132)',
-    },
-      pinIcon: { 
-          height: 40, 
-          width: 40 
-    }
-});
 
 export default Map;
