+++ src/components/styles/Styles.CalloutButton.js
@@ -1,0 +1,33 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+    container: { 
+        flex: 1 
+    },
+    image: {
+        backgroundColor: 'transparent',
+        height: 20,
+        width: 20
+    },
+    imageIcon: {
+        height: 20,
+        width: 20
+    },
+    textFind: {
+        fontSize: 15
+    },
+    searchView: { 
+        marginTop: '7%', 
+        marginLeft: '77%' 
+    },
+    findView: { 
+        alignItems: 'center', 
+        bottom: 2 
+    },
+    userView: { 
+        alignItems: 'flex-start', 
+        right: '5%' 
+    },
+});
+
+export {styles};
