+++ src/routes/Navigation.js
@@ -1,6 +1,6 @@
-import Home from '../Home';
-import Login from '../Login';
-import Map from '../Map';
+import Home from '../components/components_A/Home';
+import Login from '../components/components_A/Login';
+import Map from '../components/components_A/Map';
 import { createStackNavigator } from 'react-navigation';
 
 export default createStackNavigator({
@@ -22,4 +22,4 @@ export default createStackNavigator({
             header: null
         }
     }
-})
+});
